+++ src/components/editor-page/editor-page.tsx
@@ -32,6 +32,7 @@ import { RenderIframe } from './renderer-pane/render-iframe'
 import { Sidebar } from './sidebar/sidebar'
 import { Splitter } from './splitter/splitter'
 import { DualScrollState, ScrollState } from './synced-scroll/scroll-props'
+import { RendererType } from '../render-page/rendering-message'
 
 export interface EditorPagePathParams {
   id: string
@@ -115,13 +116,15 @@ export const EditorPage: React.FC = () => {
               showRight={ editorMode === EditorMode.PREVIEW || editorMode === EditorMode.BOTH }
               right={
                 <RenderIframe
+                  frameClasses={'h-100 w-100'}
                   markdownContent={ markdownContent }
                   onMakeScrollSource={ setRendererToScrollSource }
                   onFirstHeadingChange={ updateNoteTitleByFirstHeading }
                   onTaskCheckedChange={ SetCheckboxInMarkdownContent }
                   onFrontmatterChange={ setNoteFrontmatter }
                   onScroll={ onMarkdownRendererScroll }
-                  scrollState={ scrollState.rendererScrollState }/>
+                  scrollState={ scrollState.rendererScrollState }
+                  rendererType={ RendererType.DOCUMENT }/>
               }
               containerClassName={ 'overflow-hidden' }/>
             <Sidebar/>